Update 25.02.2022 / GF
Combining three functions into a single one which calculates the apparent magnitude in an appropriate filter
